+++ packages/mayan/src/numerals.ts
@@ -1,24 +1,24 @@
 const numbers = [
-	'\ud834\uDEE0', /** #0 𝋠 U+1D2E0  mayan numeral 0  */
-	'\ud834\uDEE1', /** #1 𝋡 U+1D2E1  mayan numeral 1  */
-	'\ud834\uDEE2', /** #2 𝋢 U+1D2E2  mayan numeral 2  */
-	'\ud834\uDEE3', /** #3 𝋣 U+1D2E3  mayan numeral 3  */
-	'\ud834\uDEE4', /** #4 𝋤 U+1D2E4  mayan numeral 4  */
-	'\ud834\uDEE5', /** #5 𝋥 U+1D2E5  mayan numeral 5  */
-	'\ud834\uDEE6', /** #6 𝋦 U+1D2E6  mayan numeral 6  */
-	'\ud834\uDEE7', /** #7 𝋧 U+1D2E7  mayan numeral 7  */
-	'\ud834\uDEE8', /** #8 𝋨 U+1D2E8  mayan numeral 8  */
-	'\ud834\uDEE9', /** #9 𝋩 U+1D2E9  mayan numeral 9  */
-	'\ud834\uDEEA', /** #10 𝋪 U+1D2EA mayan numeral 10 */
-	'\ud834\uDEEB', /** #11 𝋫 U+1D2EB mayan numeral 11 */
-	'\ud834\uDEEC', /** #12 𝋬 U+1D2EC mayan numeral 12 */
-	'\ud834\uDEED', /** #13 𝋭 U+1D2ED mayan numeral 13 */
-	'\ud834\uDEEE', /** #14 𝋮 U+1D2EE mayan numeral 14 */
-	'\ud834\uDEEF', /** #15 𝋯 U+1D2EF mayan numeral 15 */
-	'\ud834\uDEF0', /** #16 𝋰 U+1D2F0 mayan numeral 16 */
-	'\ud834\uDEF1', /** #17 𝋱 U+1D2F1 mayan numeral 17 */
-	'\ud834\uDEF2', /** #18 𝋲 U+1D2F2 mayan numeral 18 */
-	'\ud834\uDEF3', /** #19 𝋳 U+1D2F3 mayan numeral 19 */
+	'\ud834\uDEE0', /** #0 𝋠 U+1D2E0  Mayan numeral 0  */
+	'\ud834\uDEE1', /** #1 𝋡 U+1D2E1  Mayan numeral 1  */
+	'\ud834\uDEE2', /** #2 𝋢 U+1D2E2  Mayan numeral 2  */
+	'\ud834\uDEE3', /** #3 𝋣 U+1D2E3  Mayan numeral 3  */
+	'\ud834\uDEE4', /** #4 𝋤 U+1D2E4  Mayan numeral 4  */
+	'\ud834\uDEE5', /** #5 𝋥 U+1D2E5  Mayan numeral 5  */
+	'\ud834\uDEE6', /** #6 𝋦 U+1D2E6  Mayan numeral 6  */
+	'\ud834\uDEE7', /** #7 𝋧 U+1D2E7  Mayan numeral 7  */
+	'\ud834\uDEE8', /** #8 𝋨 U+1D2E8  Mayan numeral 8  */
+	'\ud834\uDEE9', /** #9 𝋩 U+1D2E9  Mayan numeral 9  */
+	'\ud834\uDEEA', /** #10 𝋪 U+1D2EA Mayan numeral 10 */
+	'\ud834\uDEEB', /** #11 𝋫 U+1D2EB Mayan numeral 11 */
+	'\ud834\uDEEC', /** #12 𝋬 U+1D2EC Mayan numeral 12 */
+	'\ud834\uDEED', /** #13 𝋭 U+1D2ED Mayan numeral 13 */
+	'\ud834\uDEEE', /** #14 𝋮 U+1D2EE Mayan numeral 14 */
+	'\ud834\uDEEF', /** #15 𝋯 U+1D2EF Mayan numeral 15 */
+	'\ud834\uDEF0', /** #16 𝋰 U+1D2F0 Mayan numeral 16 */
+	'\ud834\uDEF1', /** #17 𝋱 U+1D2F1 Mayan numeral 17 */
+	'\ud834\uDEF2', /** #18 𝋲 U+1D2F2 Mayan numeral 18 */
+	'\ud834\uDEF3', /** #19 𝋳 U+1D2F3 Mayan numeral 19 */
 ] as const
 
 export const Ma = {