+++ packages/hieroglyphic/CHANGELOG.md
@@ -1,4 +1,4 @@
-# hieroglyphic Changelog
+# Hieroglyphic Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
 ## [0.0.2]  2024-05-01
@@ -8,4 +8,4 @@
 
 ## [0.0.1]  2024-04-30
 ### Added
-- Initial project covert number into hieroglyphic numerals
+- Initial project covert number into Hieroglyphic numerals