+++ packages/mayan/CHANGELOG.md
@@ -1,6 +1,6 @@
-# mayan Changelog
+# Mayan Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
 ## [0.0.1]  2024-04-25
 ### Added
-- Initial project covert number into mayan numerals
+- Initial project covert number into Mayan numerals